+++ HISTORY.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* Allow user to specify thread count
+
+## 1.0.1
+
+* fix FileNotFoundError when unlinking shared memory
+
 ## 1.0.0
 
 * it works